+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [0.5.3] - 2025-03-19
 ### Added
 - Add prop `hgtReadFileRate` to `MapContainerModule`. To control `HgtReader` read file throttle rate.
 
@@ -107,7 +107,7 @@ Just updated README.md
 ## [0.0.1] - 2024-10-25
 First bumpy version
 
-[Unreleased]: https://github.com/jhotadhari/react-native-mapsforge-vtm/compare/v0.5.2...HEAD
+[0.5.3]: https://github.com/jhotadhari/react-native-mapsforge-vtm/compare/v0.5.2...v0.5.3
 [0.5.2]: https://github.com/jhotadhari/react-native-mapsforge-vtm/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/jhotadhari/react-native-mapsforge-vtm/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/jhotadhari/react-native-mapsforge-vtm/compare/v0.4.0...v0.5.0